requirements-lock.txt
absl-py==2.3.1
albumentations==1.3.1
astunparse==1.6.3
blinker==1.9.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
Flask==2.3.2
Flask-Cors==4.0.0
flatbuffers==25.9.23
fonttools==4.60.1
gast==0.6.0
google-pasta==0.2.0
grpcio==1.76.0
gunicorn==21.2.0
h5py==3.15.1
idna==3.11
imageio==2.37.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.2
keras==3.12.0
kiwisolver==1.4.9
lazy_loader==0.4
libclang==18.1.1
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.7.2
mdurl==0.1.2
ml_dtypes==0.5.3
namex==0.1.0
networkx==3.5
numpy==1.26.4
opencv-python==4.8.0.74
opencv-python-headless==4.11.0.86
opt_einsum==3.4.0
optree==0.17.0
packaging==25.0
pandas==2.0.3
Pillow==10.0.0
protobuf==6.33.0
Pygments==2.19.2
pyparsing==3.0.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.2
PyYAML==6.0.3
qudida==0.0.4
requests==2.31.0
rich==14.2.0
scikit-image==0.24.0
scikit-learn==1.3.0
scipy==1.16.3
seaborn==0.12.2
six==1.17.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.2.0
threadpoolctl==3.6.0
tifffile==2025.10.16
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
Werkzeug==2.3.6
wrapt==2.0.0
